refactor: use createSlice API
