+++ src/features/counter.ts
@@ -1,0 +1,23 @@
+import { createSlice, PayloadAction } from "@reduxjs/toolkit";
+
+export type CounterState = { count: number };
+export const initialState: CounterState = { count: 0 };
+
+export const counterSlice = createSlice({
+  name: "counter",
+  initialState,
+  reducers: {
+    added: (state, action: PayloadAction<number>) => ({
+      ...state,
+      count: state.count + action.payload,
+    }),
+    decremented: (state) => ({
+      ...state,
+      count: state.count - 1,
+    }),
+    incremented: (state) => ({
+      ...state,
+      count: state.count + 1,
+    }),
+  },
+});
