+++ src/index.tsx
@@ -3,7 +3,8 @@ import ReactDOM from "react-dom/client";
 import { Provider } from "react-redux";
 import { configureStore } from "@reduxjs/toolkit";
 import reportWebVitals from "./reportWebVitals";
-import { counterReducer } from "./features/counter/counter-reducer";
+// import { counterReducer } from "./features/counter/counter-reducer";
+import { counterSlice } from "./features/counter";
 import App from "./App";
 import "./index.css";
 import "semantic-ui-css/semantic.min.css";
@@ -11,7 +12,7 @@ import "semantic-ui-css/semantic.min.css";
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
 );
-const store = configureStore({ reducer: counterReducer });
+const store = configureStore({ reducer: counterSlice.reducer });
 
 root.render(
   <React.StrictMode>
